Add method for determining topLevelDesinations in AppBarConfiguration
Added utility public method `isTopLevelDestination` to determine whether a given destination is a topLevelDestination. It replaces `matchDestinations` which didn't differentiate between graph ids and startDestinationId of graphs when the destination could be a nested graph, such as a MenuItem NavGraph within a Menu.

Test: /.gradlew navigation:navigation-ui:cC
Bug: 238496771
Relnote: "Behavior change: When passing the ID of a navigation graph to `AppBarConfiguration` (such as via a `Menu`), `NavigationUI` now only considers the start destination of that navigation graph as a top level destination, rather than incorrectly marking every destination within the graph as a top level destination. The behavior of passing the ID of an individual destination is unchanged. This same functionality is available to your own code via the new `isTopLevelDestination` method on `AppBarConfiguration`."

Change-Id: Ie936ece98a80c2a13fbfb77b83a0ba86e926ebd9