Create an XTypeName for XProcessing and XPoet
Create a xprocessing type name that wraps JavaPoet and KotlinPoet type names. The KotlinPoet name will only be available in KSP, for Javac and KAPT the name will be UNAVAILABLE, the reasoning is that only in KSP will Kotlin code will be generated as it contains the best type information.

* XPoet APIs are migrated to take XTypeName instead of JavaPoet type name.
* For now KSP KotlinPoet name building logic is copied from JavaPoet's version.

Migration and deprecation of JavaPoet as API in xprocessing will be done in follow-up CLs along with any test that verifies JavaPoet name will also verify KotlinPoet name.

Bug: 127483380
Test: XArrayType and a bit in XTypeElement
Change-Id: I76f0afbdb2a2fc6fa50c930ccc0fd7d50b998197