Add @Keep to CarAudioRecord classes
Previously we found issues on release builds where the VoiceAccess
feature would crash on both AAP and AAOS. The root cause is that the
AutomotiveCarAudioRecord and ProjectedCarAudioRecord classes are
obfuscated in the compiling process so the client couldn't find them. To
fix this, we need to add @Keep to both classes.

Test: Manual testing
Relnote: Add @Keep to CarAudioRecord classes
Bug: 247614190
Change-Id: I19f6fd277a275b95cbd3d239ff4f22cd4ef1bb4f