Add feature detection using WebView Manifest entry
Till today, webkit feature detection was dependant on WebView being
loaded into the calling process. This would not work for
ProcessGlobalConfig since it needs to check for features prior to
loading WebView. One naive way of solving this would be to hard code the
version codes of the WebView binaries in which a particular feature was
added. This approach has very poor readability and is error-prone.

Instead, this CL provides a framework which allows features to be added
to the WebView manifest that can be queried by the webkit library
without loading webview.

This CL has a corresponding CL in chromium which adds the entry to the
Android manifest: https://crrev.com/c/3921574

Test: Locally tested using ProcessGlobalConfigActivity
Change-Id: I332c244cea3777552e025e22a5f659bcad430ca3