Start generating Kotlin for FieldReadWriteWriter.readFromCursor()
To enable partial migration to XPoet this CL makes it so that the Pojo cursor reading code is generated in a separate Kotlin class that is then called from the Java Dao impl. This partial migration moves PrimitiveColumnTypeAdapter to XPoet, other type adapters will follow.

Additionally, this CL moves various JavaPoet TypeName usages in Room's value objects to XTypeName, mainly Pojo.typeName, most of the move is a refactor using toJavaPoet() that is a temporary compatibility API to be removed.

Due to the amount of type converters used in the integration test app this change only validates that the Java runtime behaves as expected and that the Kotlin generated code compiles (via KotlinCodeGenTest) but does not validate the runtime behaviour of the Kotlin codegen. We need to have all type converters and cursor readers migrated to run the integration tests.

As more of Room is migrated more XPoet APIs are added, in this CL we now have:
* begin, next and end control flow
* code block of a new instance, using the 'new' keyword in Java, but not in Kotlin
* code block of an unsafe cast, using the parenthesis syntax in Java and the 'as' keyword in Kotlin
* addLocalVariable is fixed (now that is being used)

Test: Existing tests + KotlinCodeGenTest
Change-Id: Id25a6e874fe9a03768ceca7e2c2a606ac2f469a2