Changing the handling of undefined default value in columns.
If the default value is not defined, it should be noted as something other than 'null', since this clashes with the case where a default value is actually a null value and the user cannot differentiate between the two while debugging. This has been changed to print "undefined" in this case. Other alternatives that could be used include "missing" or "notDefined".

Bug: 187865510
Test: TableInfoTest.java
Change-Id: I504e3a390c1b88b53111b878f0a51358a1fa2d83