Update CompareAspectRatiosByDistanceToTargetRatio to compare by overlapping area
Added additional full FOV ratio parameter so that we can check whether mapping area of the checking ratio can cover the mapping area of the target aspect ratio in the full active array ratio space. And we can also check the overlapping area size in the full active array ratio space to know which ratio is closer to the target aspect ratio value.

Rename CompareAspectRatiosByDistanceToTargetRatio to CompareAspectRatiosByMappingAreaInFullFovAspectRatioSpace to match its new algorithm

Bug: 247471840
Test: AspectRationUtilTest && SupportedSurfaceCombinationTest

Change-Id: Ifc45e8e9537a177cf9ef2bcc8d4f226434ea6950