Replace usage of ConcurrentHashMap in MediaControllerCompat
When using android_sdk_version < 21,java.lang.NoClassDefFoundError: j$.util.concurrent.ConcurrentHashMap is encountered. The use of ConcurrentHashMap in combination with Java 8 Library Desugaring has the limitation that ConcurrentHashMap cannot be used until the additional DEX files are loaded. This is mitigated by using a synchronised set, as mRegisteredCallbacks is only used to store unique callbacks.

Bug: b/250903905
Test: ./media/version-compat-tests/runtest.sh 4

Change-Id: I06a3549c4828288dd0289ab4be2717ddcf7d1e2d