Support for generating baseline profiles without root (API33+)
Fixes:250083467
Fixes:253094958
Test: BaselineProfileRuleTest # on raven running T / API 33, user build
Test: ./gradlew benchmark:b-m:cC

Relnote: """
// BENCHMARK RELEASE NOTES

BaselineProfileRule no longer requires root on Android 13 (API 33),
and is no longer experimental.

This change also fixes how profiles from an app are flushed to disk on
unrooted devices, but requires updating the target app's
profileinstaller dependency.

To use `BaselineProfileRule` and `CompilationMode.Partial()` on an
unrooted device, you must also update your target app to use
`androidx.profileinstaller.profileinstaller:1.3.0-alpha01`. This
enables flushing the profile to disk correctly, so that it can be
compiled/extracted.

// PROFILE INSTALLER RELEASE NOTES

Adds a new shell broadcast that enables Macrobenchmark to fully flush
in-memory profile data to disk, to be included in baseline profile
generation. This is required to use the macrobenchmark library to
capture baseline profiles with `BaselineProfileRule`, and evaluate
profile performance using `CompilationMode.Partial()`.
"""

Change-Id: Ie0a7d13cbe34476113bad955e47ed771d84e65a4