ScalingLazyColumn clarify handling of viewport center line and list item center line calculation.
Ensure that a consistent approach is applied to the rounding of center lines for viewport and list items.
Adjust the top and bottom padding calculations based on the new rounding approach.
If the viewport or a list item is an odd number of pixels in size then its center line will now always be rounded down.

Bug: 253183722
Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Change-Id: Icebef25e0e083ae298cf0c81b1c07fb277f165f4