SurfaceRequest new method API review
This CL add an invalidate() method to notify the surface requester that
the previously provided surface is no longer valid and a new
SurfaceRequest is needed for obtaining the new Surface.

For details, please see design doc: go/camerax-surface-view-viewpager2.

Relnote: "Add invalidate() method to SurfaceRequest. The Surface
provider can notify that the previously provided surface is no longer
valid."

Bug: 254192793
Test: manual test and ./gradlew bOS
Change-Id: Idc649eabf476d7f667d5aa613f6be03f1815d2e8