Always retry the in-flight request when paused
When the camera resets, camera-camera2 may not send error callback to camera-core. This has been causing the UseCaseCombinationTest to fail on LEGACY devices. We cannot rely solely on onCaptureFailure() to retry the request.

In this CL, we always retry when the manager is paused, which is consistent with the old pipeline's behavior.

Bug: 255737797
Test: manual test and ./gradlew bOS
Change-Id: I5129326aeea1e05923cbd83ba8154a9e2bc2e17e