Internal change from using SemanticsSort to 'setTraversal'
This change is meant to be pre-work for a larger refactoring of focus order in Compose and shouldn’t introduce any changes in A11y behavior.

Issues seen in b/227206751 and https://github.com/maxim-pandra/ComplexAccessibility are part of an problem where parent/child relations in existing Semantics nodes block some ways of ordering—for example, if there were a row containing multiple text elements, then a following button overlaid on the row, TalkBack would read through all text elements before jumping back up the screen to the button. This is because `SemanticsSort` will always read the children of a parent node before moving on. A comparison independent of parent/child hierarchies would read the text children up to where the button element is on the screen, read the button element, and then proceed with the rest of the text children.

This patch is just meant to check that we can use `setTraversalBefore` to determine our own custom order for semantics nodes (overriding the existing hierarchies). `setTraversalValues` in `AndroidComposeViewAccessibilityDelegateCompat` should order semantics nodes in the same way `SemanticsSort` did, except by setting the `traversalBefore` value for each node that will be read by TalkBack.

This ordering is set in a mapping when semantic nodes are first retrieved. Then when a nodeInfo is created, the code now looks up our custom traversal order and uses `setTraversalBefore` on that ANI. That way, when TalkBack goes through the ANI tree, it’ll use the `traversalBefore` parameter and traverse in the way that we want regardless, of parent/child hierarchies.

Bug: 186443263
Test: compose/ui/AndroidAccessibilityTest/testCreateAccessibilityNodeInfo_forTraversalOrder

Change-Id: I339010bea62afb359e63d6e63889acff2c72cf27