Fix XType.supertypes in KSP.
XTypes.supertypes should return the resolved supertypes. For example,
`Foo<Bar>` supertype should be `FooBase<Bar>` rather than `FooBase<T>`.

This CL fixes this for KSP types, as this has always worked properly for
javac types.

This CL also fixes a few minor issues to match javac as described below.

This CL updates the description of `XTypes#supertypes()` to include the
ordering (class then interfaces) to match that of the javac API for
`Types#directSupertypes()` from which this method was derived:

https://docs.oracle.com/javase/8/docs/api/javax/lang/model/util/Types.html#directSupertypes-javax.lang.model.type.TypeMirror-

This CL also adds `Object` as the super class if there are no explicit
superclasses declared on the interface/class to match the behavior of
javac.

This CL also fixes `XTypes#isError()` which should be disjoint from
`XTypes#isJavacWildcardType()` (i.e. if its a wildcard type it shouldn't
also be an error type).

Test: XTypeTest
Change-Id: I26bf8d22df09f32d2cd4d04c02bab0f71f824a87