Make some input prediction library members final
As pointed out in the original code review (aosp/2242040), some
members can be made final. Other than the ones pointed out in
that review, I checked the rest of the files and applied it where
possible. It also has a slight change of logic in the close logic
to free up the predictor resources, and a small documentation
correction.

Bug: 232941452
Test: Compiles successfully
Change-Id: I3fab1ca864250c23d76c3a1e3ca8a1f6abbbb1c3