Set SwipeableV2 density from layout modifier
Instead of obtaining the density from composition and passing it as a constructor param, it is now a (mutable) property on SwipeableV2State.
We prefer this over using a composed modifier to obtain and update the density as it reduces overhead. It also reduces the burden on SwipeableV2State constructor callsites as creating the state doesn't require obtaining the density by themselves anymore.

Test: Existing tests pass
Relnote: N/A

Change-Id: I895c41a66e852cc950536c422e3ba7dcb01d6a3f