Cancel graph state collection after UseCaseCamera is closed
Previously, we directly cancel the job that collects the graph state
from CameraPipe. This causes a race condition during shutdown where
UseCaseCamera would never receive GraphStateStopped, and thereby causing
the CameraState to not transition to CLOSED.

Here we add a `closed` variable, and we cancel the coroutine scope only
when the UseCaseCamera is closed and we've received GraphStateStopped
from CameraPipe.

Bug: None
Test: androidx.camera.view.PreviewViewStreamStateTest
Change-Id: I11caa175fa1d8572426743c046722d26c49bbafd