Fix experimental annotation on TestMonotonicFrameClock.
ExperimentalCoroutinesApi is the annotation used to mark kotlinx
coroutines apis that are experimental. We have our own annotation,
ExperimentalTestApi, for compose test APIs that are experimental. This
change uses the latter annotation on our APIs, and opts-in to the former
where necessary to consume experimental coroutines APIs.

Test: n/a, only changing annotations
Relnote: "TestMonotonicFrameClock now uses correct experimental
    annotation."
Change-Id: I95c9efb389cad81cb9182ff83910f53c7a99c239