Stop process output buffer after MediaCodec is reset
It could cause an error when the output frame
buffer is processed/released after the MediaCodec
is reset.

In EncoderImpl, the onOutputBufferAvailable event
will be posted to the mEncoderExecutor, it could
make the callback be processed after the MediaCodec
is reset.

This CL add an extra stopped flag to the MediaCodec.Callback
which will be used to ignore the further processing
after flush/reset the MediaCodec.

Bug: 259358028
Test: VideoRecordingTest
Change-Id: I2a91ec21603e74a99932b4439933b082ac4f7fa9