Fix initializationError on Android API 17
The reason for the failure is when Parameterized
tests require iterating the tests, and the tests include
objects not available on Andorid 17.

This CL creates a TestRunner wrapper to filter the tests by
Android API level. This TestRunner will ignore all the
internal test classes when android API level of the DUT is
not in the range of target SDK.

Bug: 260086103
Test: ViewPortTest on Android API level 17 emulator
Change-Id: I14d792b73db44c2444327cb75a84f566a814944c